Feature/delegate proposal (#32)
* bump telegram-bot lib to 13.x, improve lcd_url loading and exception readability

* refactoring - consts and files structure

* add __init__ files

* add is_syncing method

* add reading env variables

* add sentry sync check jobs, refactoring

* update docs

* move some messages

* print -> log

* removing global jobs bug - hotfix

* hotfix - tests

* add voting with website, remove mnemonic, refactor requirements.txt

* after merge fix

+++ bot/bot.py
@@ -1,8 +1,8 @@
 import atexit
 import subprocess
 
-from telegram import TelegramError
 from telegram.ext import Updater, PicklePersistence, CommandHandler, CallbackQueryHandler, MessageHandler, Filters
+from telegram import TelegramError
 
 from constants.constants import *
 from constants.messages import BOT_STARTUP_LOG, BOT_RESTARTED
@@ -73,9 +73,8 @@ def setup_existing_user(dispatcher):
 
         # Somehow session.data does not get updated if all users block the bot.
         # That's why we delete the file ourselves.
-        if len(dispatcher.persistence.user_data) == 0:
-            if os.path.exists(session_data_path):
-                os.remove(session_data_path)
+        if len(dispatcher.persistence.user_data) == 0 and os.path.exists(session_data_path):
+            os.remove(session_data_path)
 
 
 def main():
@@ -110,7 +109,7 @@ def main():
     LCD endpoint: {LCD_ENDPOINT}
     Sentry nodes: {SENTRY_NODES}
     Node IP: {NODE_IP}
-    MNEMONIC: {"SET" if MNEMONIC else "NOT SET"}
+    WHITELISTED USERS: {ALLOWED_USER_IDS}
     ==========================================================================
     ==========================================================================
     """)
